+++ src/Driver/Pdo/AbstractPdoCommand.php
@@ -211,7 +211,7 @@ abstract class AbstractPdoCommand extends AbstractCommand implements PdoCommandI
                 }
                 break;
             } catch (PDOException $e) {
-                $rawSql = $rawSql ?: $this->getRawSql();
+                $rawSql ??= $this->getRawSql();
                 $e = (new ConvertException($e, $rawSql))->run();
 
                 if ($this->retryHandler === null || !($this->retryHandler)($e, $attempt)) {
@@ -260,19 +260,26 @@ abstract class AbstractPdoCommand extends AbstractCommand implements PdoCommandI
      */
     protected function logQuery(string $rawSql, string $category): void
     {
-        $this->logger?->log(LogLevel::INFO, $rawSql, [$category]);
+        $this->logger?->log(LogLevel::INFO, $rawSql, [$category, 'type' => LogType::QUERY]);
     }
 
     protected function queryInternal(int $queryMode): mixed
     {
-        $rawSql = $this->getRawSql();
         $logCategory = self::class . '::' . $this->getQueryMode($queryMode);
 
-        $this->logQuery($rawSql, $logCategory);
+        if ($this->logger !== null) {
+            $rawSql = $this->getRawSql();
+            $this->logQuery($rawSql, $logCategory);
+        }
 
         $queryContext = new CommandContext(__METHOD__, $logCategory, $this->getSql(), $this->getParams());
 
-        $this->profiler?->begin($rawSql, $queryContext);
+        /**
+         * @psalm-var string $rawSql
+         * @psalm-suppress RedundantConditionGivenDocblockType
+         * @psalm-suppress DocblockTypeContradiction
+         */
+        $this->profiler?->begin($rawSql ??= $this->getRawSql(), $queryContext);
         try {
             /** @psalm-var mixed $result */
             $result = parent::queryInternal($queryMode);