+++ src/Driver/Pdo/LogType.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Db\Driver\Pdo;
+
+final class LogType
+{
+    public const CONNECTION = 'connection';
+    public const QUERY = 'query';
+    public const TRANSACTION = 'transaction';
+}