+++ tests/AbstractQueryBuilderTest.php
@@ -212,14 +212,21 @@ abstract class AbstractQueryBuilderTest extends TestCase
      *
      * @psalm-param array<array-key, string> $columns
      */
-    public function testBatchInsert(string $table, array $columns, iterable $rows, string $expected): void
-    {
-        $db = $this->getConnection();
-
+    public function testBatchInsert(
+        string $table,
+        array $columns,
+        iterable $rows,
+        string $expected,
+        array $expectedParams = [],
+    ): void {
+        $db = $this->getConnection(true);
         $qb = $db->getQueryBuilder();
-        $sql = $qb->batchInsert($table, $columns, $rows);
+
+        $params = [];
+        $sql = $qb->batchInsert($table, $columns, $rows, $params);
 
         $this->assertSame($expected, $sql);
+        $this->assertSame($expectedParams, $params);
     }
 
     /**
@@ -241,10 +248,12 @@ abstract class AbstractQueryBuilderTest extends TestCase
 
         [$sql, $params] = $db->getQueryBuilder()->build($query);
 
-        $replacedQuotes = DbHelper::replaceQuotes((string) $expected, $db->getDriverName());
-
-        $this->assertIsString($replacedQuotes);
-        $this->assertEquals('SELECT *' . (empty($expected) ? '' : ' WHERE ' . $replacedQuotes), $sql);
+        $this->assertEquals(
+            'SELECT *'
+            . ($db->getDriverName() === 'oci' ? ' FROM DUAL' : '')
+            . (empty($expected) ? '' : ' WHERE ' . DbHelper::replaceQuotes($expected, $db->getDriverName())),
+            $sql
+        );
         $this->assertEquals($expectedParams, $params);
     }
 
@@ -344,9 +353,9 @@ abstract class AbstractQueryBuilderTest extends TestCase
         [$sql, $params] = $qb->build($query);
 
         $this->assertSame(
-            'SELECT *' . (
-                empty($expected) ? '' : ' WHERE ' . DbHelper::replaceQuotes($expected, $db->getDriverName())
-            ),
+            'SELECT *'
+            . ($db->getDriverName() === 'oci' ? ' FROM DUAL' : '')
+            . (empty($expected) ? '' : ' WHERE ' . DbHelper::replaceQuotes($expected, $db->getDriverName())),
             $sql,
         );
         $this->assertSame($expectedParams, $params);
@@ -467,10 +476,12 @@ abstract class AbstractQueryBuilderTest extends TestCase
 
         [$sql, $params] = $db->getQueryBuilder()->build($query);
 
-        $replacedQuotes = DbHelper::replaceQuotes($expected, $db->getDriverName());
-
-        $this->assertIsString($replacedQuotes);
-        $this->assertSame('SELECT *' . (empty($expected) ? '' : ' WHERE ' . $replacedQuotes), $sql);
+        $this->assertSame(
+            'SELECT *'
+            . ($db->getDriverName() === 'oci' ? ' FROM DUAL' : '')
+            . (empty($expected) ? '' : ' WHERE ' . DbHelper::replaceQuotes($expected, $db->getDriverName())),
+            $sql
+        );
         $this->assertSame($expectedParams, $params);
     }
 
@@ -1301,12 +1312,13 @@ abstract class AbstractQueryBuilderTest extends TestCase
 
         [$sql, $params] = $qb->build($query);
 
-        $this->assertSame(
-            <<<SQL
-            SELECT DISTINCT *
-            SQL,
-            $sql,
-        );
+        $expected = 'SELECT DISTINCT *';
+
+        if ($this->getDriverName() === 'oci') {
+            $expected .= ' FROM DUAL';
+        }
+
+        $this->assertSame($expected, $sql);
         $this->assertSame([], $params);
     }
 
@@ -1553,16 +1565,16 @@ abstract class AbstractQueryBuilderTest extends TestCase
     public function testCreateView(): void
     {
         $db = $this->getConnection();
-
         $qb = $db->getQueryBuilder();
 
+        $expected = 'CREATE VIEW [[animal_view]] AS SELECT [[1]]';
+
+        if ($this->getDriverName() === 'oci') {
+            $expected .= ' FROM DUAL';
+        }
+
         $this->assertSame(
-            DbHelper::replaceQuotes(
-                <<<SQL
-                CREATE VIEW [[animal_view]] AS SELECT [[1]]
-                SQL,
-                $db->getDriverName(),
-            ),
+            DbHelper::replaceQuotes($expected, $db->getDriverName()),
             $qb->createView('animal_view', (new query($db))->select('1')),
         );
     }