+++ tests/Provider/QueryBuilderProvider.php
@@ -146,7 +146,7 @@ class QueryBuilderProvider
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => 'test@example.com', ':qp1' => 'silverfire', ':qp2' => 'Kyiv {{city}}, Ukraine'],
+                'expectedParams' => [':qp0' => 'test@example.com', ':qp1' => 'silverfire', ':qp2' => 'Kyiv {{city}}, Ukraine'],
             ],
             'escape-danger-chars' => [
                 'customer',
@@ -158,7 +158,7 @@ class QueryBuilderProvider
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => "SQL-danger chars are escaped: '); --"],
+                'expectedParams' => [':qp0' => "SQL-danger chars are escaped: '); --"],
             ],
             'customer2' => [
                 'customer',
@@ -172,11 +172,11 @@ class QueryBuilderProvider
                 [['no columns passed']],
                 'expected' => DbHelper::replaceQuotes(
                     <<<SQL
-                    INSERT INTO [[customer]] () VALUES (:qp0)
+                    INSERT INTO [[customer]] VALUES (:qp0)
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => 'no columns passed'],
+                'expectedParams' => [':qp0' => 'no columns passed'],
             ],
             'bool-false, bool2-null' => [
                 'type',
@@ -188,7 +188,7 @@ class QueryBuilderProvider
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => 0, ':qp1' => null],
+                'expectedParams' => [':qp0' => false, ':qp1' => null],
             ],
             'wrong' => [
                 '{{%type}}',
@@ -200,7 +200,7 @@ class QueryBuilderProvider
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => null, ':qp1' => null],
+                'expectedParams' => [':qp0' => null, ':qp1' => null],
             ],
             'bool-false, time-now()' => [
                 '{{%type}}',
@@ -212,7 +212,7 @@ class QueryBuilderProvider
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => null],
+                'expectedParams' => [':qp0' => false],
             ],
             'column table names are not checked' => [
                 '{{%type}}',
@@ -224,7 +224,7 @@ class QueryBuilderProvider
                     SQL,
                     static::$driverName,
                 ),
-                [':qp0' => null, ':qp1' => null],
+                'expectedParams' => [':qp0' => true, ':qp1' => false],
             ],
             'empty-sql' => [
                 '{{%type}}',
@@ -235,6 +235,23 @@ class QueryBuilderProvider
                     }
                 })(),
                 '',
+            ],
+            'empty columns and non-exists table' => [
+                'non_exists_table',
+                [],
+                'values' => [['1.0', '2', 10, 1]],
+                'expected' => DbHelper::replaceQuotes(
+                    <<<SQL
+                    INSERT INTO [[non_exists_table]] VALUES (:qp0, :qp1, :qp2, :qp3)
+                    SQL,
+                    static::$driverName,
+                ),
+                'expectedParams' => [
+                    ':qp0' => '1.0',
+                    ':qp1' => '2',
+                    ':qp2' => 10,
+                    ':qp3' => 1,
+                ],
             ],
         ];
     }