+++ tests/Support/DbHelper.php
@@ -21,7 +21,7 @@ final class DbHelper
 {
     public static function changeSqlForOracleBatchInsert(string &$str): void
     {
-        $str = str_replace('INSERT INTO', 'INSERT ALL  INTO', $str) . ' SELECT 1 FROM SYS.DUAL';
+        $str = str_replace('INSERT INTO', 'INSERT ALL INTO', $str) . ' SELECT 1 FROM SYS.DUAL';
     }
 
     public static function getPsrCache(): CacheInterface